requirements.txt
-i https://pypi.org/simple
aiodns==3.0.0 ; python_full_version >= '3.5.2'
aiohttp==3.8.5 ; python_full_version >= '3.5.2'
aiosignal==1.3.1 ; python_version >= '3.7'
async-timeout==4.0.2 ; python_version >= '3.6'
attrs==23.1.0 ; python_version >= '3.7'
ccxt==4.0.39
certifi==2023.7.22 ; python_version >= '3.6'
cffi==1.15.1
cfgv==3.3.1 ; python_full_version >= '3.6.1'
charset-normalizer==3.2.0 ; python_full_version >= '3.7.0'
cryptography==41.0.2 ; python_version >= '3.7'
distlib==0.3.7
exceptiongroup==1.1.2 ; python_version < '3.11'
filelock==3.12.2 ; python_version >= '3.7'
frozenlist==1.4.0 ; python_version >= '3.8'
identify==2.5.26 ; python_version >= '3.8'
idna==3.4 ; python_version >= '3.5'
iniconfig==2.0.0 ; python_version >= '3.7'
multidict==6.0.4 ; python_version >= '3.7'
nodeenv==1.8.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
numpy==1.25.1
packaging==23.1 ; python_version >= '3.7'
pandas==2.0.3
platformdirs==3.9.1 ; python_version >= '3.7'
pluggy==1.2.0 ; python_version >= '3.7'
pre-commit==3.3.3
pycares==4.3.0
pycparser==2.21
pytest==7.4.0
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0
pytz==2023.3
pyyaml==6.0.1
requests==2.31.0 ; python_version >= '3.7'
scipy==1.11.1
setuptools==68.0.0 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tomli==2.0.1 ; python_version < '3.11'
tqdm==4.65.0
tzdata==2023.3 ; python_version >= '2'
urllib3==2.0.4 ; python_version >= '3.7'
virtualenv==20.24.2 ; python_version >= '3.7'
yarl==1.9.2 ; python_full_version >= '3.5.2'
